requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
httpx==0.25.2
websockets==12.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
aiosqlite==0.19.0
sqlalchemy==2.0.23
bcrypt==4.1.2
